requirements.txt
django
django-filter
djangorestframework
django-cors-headers
djangorestframework-simplejwt
django-rest-passwordreset
rest-framework-roles